HMAN-177 To be consistent with other end points
