+++ src/main/java/uk/gov/hmcts/reform/hmc/controllers/LinkHearingGroupController.java
@@ -4,6 +4,8 @@ import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -59,5 +61,18 @@ public class LinkHearingGroupController {
     public void updateHearing(@RequestParam("id") String requestId,
                               @RequestBody @Valid HearingLinkGroupRequest hearingLinkGroupRequest) {
         linkedHearingGroupService.updateLinkHearing(requestId, hearingLinkGroupRequest);
+    }
+
+    @DeleteMapping(path = "/linkedHearingGroup/{id}", consumes = APPLICATION_JSON_VALUE,
+        produces = APPLICATION_JSON_VALUE)
+    @ResponseStatus(HttpStatus.OK)
+    @ApiResponses(value = {
+        @ApiResponse(code = 200, message = "Hearing group deletion processed"),
+        @ApiResponse(code = 400, message = "Invalid hearing group details found"),
+        @ApiResponse(code = 404, message = "Hearing Group id not found"),
+        @ApiResponse(code = 500, message = "Error occurred on the server")
+    })
+    public void deleteHearingGroup(@PathVariable("id") Long hearingGroupId) {
+        linkedHearingGroupService.deleteLinkedHearingGroup(hearingGroupId);
     }
 }
